% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection_rules.R
\name{incremental}
\alias{incremental}
\title{routine coverage projection rule - annual incremental to certain threshold}
\usage{
incremental(d, year_from, year_to, step, cap = 0.95)
}
\arguments{
\item{d}{a data frame containing at least (year, coverage)}

\item{year_from}{year from}

\item{year_to}{year to}

\item{step}{incremental level, e.g. 1% annual increase}

\item{cap}{threshold capping coverage, this will be replaced by historical peak in d if applicable}
}
\description{
routine coverage projection rule - annual incremental to certain threshold
}
