% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection_rules.R
\name{keep_levels}
\alias{keep_levels}
\title{routine coverage projection rule - keep future coverage at certain level}
\usage{
keep_levels(d, year_from, year_to, level)
}
\arguments{
\item{d}{a data frame containing at least (year, coverage)}

\item{year_from}{year from}

\item{year_to}{year to}

\item{level}{coverage level}
}
\description{
routine coverage projection rule - keep future coverage at certain level
}
