% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection_rules.R
\name{non_linear_scale_up}
\alias{non_linear_scale_up}
\title{routine coverage projection rule -  ia2030 non-linear scale-up
this is specifically for routine intro (no historical data, i.e. nrow(d) == 0), but can be used flexibly}
\usage{
non_linear_scale_up(
  d,
  year_from,
  year_to,
  endpoint,
  gender = NULL,
  age_from = NULL,
  age_to = NULL
)
}
\arguments{
\item{d}{a data frame containing at least (year, coverage)}

\item{year_from}{year from}

\item{year_to}{year to}

\item{endpoint}{coverage level in year_to}

\item{gender}{1 for both, 2 for males, 3 for females}

\item{age_from}{age_from}

\item{age_to}{age_to}
}
\description{
routine coverage projection rule -  ia2030 non-linear scale-up
this is specifically for routine intro (no historical data, i.e. nrow(d) == 0), but can be used flexibly
}
