% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection_rules.R
\name{sia_catch_up}
\alias{sia_catch_up}
\title{campaign coverage projection rule - one-off catch-up campaigns, specifically initial catch-up and mini-catch-up that targets missing cohorts}
\usage{
sia_catch_up(
  d,
  dat,
  vaccine_base,
  year_current,
  sia_level,
  age_from,
  age_to,
  gender
)
}
\arguments{
\item{d}{a data frame containing at least (year, coverage)}

\item{dat}{data frame containing at least vaccine_base upto year_to}

\item{vaccine_base}{vaccine to be dependent on}

\item{year_current}{year current}

\item{sia_level}{campaign coverage level}

\item{age_from}{campaign target age from}

\item{age_to}{campaign target age to}

\item{gender}{1 for both, 2 for males, 3 for females}
}
\description{
campaign coverage projection rule - one-off catch-up campaigns, specifically initial catch-up and mini-catch-up that targets missing cohorts
}
