% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection_rules.R
\name{sia_follow_up}
\alias{sia_follow_up}
\title{campaign coverage projection rule - future follow-up campaigns
this function is designed under WHO M/MR guidance
depending on the most recent campaign, routine introduction date and routine coverage level of vaccine_base}
\usage{
sia_follow_up(
  d,
  dat,
  vaccine_base,
  year_current,
  year_to,
  look_back = 4,
  sia_level = 0.9,
  age_from = 1,
  age_to = 5
)
}
\arguments{
\item{d}{a data frame containing at least (year, coverage)}

\item{dat}{data frame containing at least vaccine_base upto year_to}

\item{vaccine_base}{vaccine to be dependent on}

\item{year_current}{year current}

\item{year_to}{year to}

\item{look_back}{evaluate follow-up sia frequency from year_current - look_back}

\item{sia_level}{campaign coverage level}

\item{age_from}{campaign target age from}

\item{age_to}{campaign target age to}
}
\description{
campaign coverage projection rule - future follow-up campaigns
this function is designed under WHO M/MR guidance
depending on the most recent campaign, routine introduction date and routine coverage level of vaccine_base
}
