% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection_rules.R
\name{sia_recurrent}
\alias{sia_recurrent}
\title{campaign coverage projection rule - recurrent campaigns, e.g. cholera}
\usage{
sia_recurrent(
  d,
  dat = NULL,
  year_from,
  year_to,
  frequency,
  sia_level,
  age_from,
  age_to,
  gender
)
}
\arguments{
\item{d}{a data frame containing at least (year, coverage)}

\item{dat}{Not used. It is just here for maintaining consistent campaign projection structure}

\item{year_from}{year from}

\item{year_to}{year to}

\item{frequency}{sia frequency}

\item{sia_level}{campaign coverage level}

\item{age_from}{campaign target age from}

\item{age_to}{campaign target age to}

\item{gender}{1 for both, 2 for males, 3 for females}
}
\description{
campaign coverage projection rule - recurrent campaigns, e.g. cholera
}
