% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{vac_sce}
\alias{vac_sce}
\title{Generate Vaccination Scenario}
\usage{
vac_sce(
  input,
  key_cols = c("region", "vaccine", "activity_type", "year", "age_from", "age_to",
    "gender", "target", "coverage", "proportion_risk")
)
}
\arguments{
\item{input}{for details of setting up input, see input_check()}

\item{key_cols}{Key columns in coverage data to maintain
by default key cols in coverage data include c("region", "vaccine", "activity_type", "year", "age_from", "age_to", "gender", "target", "coverage", "proportion_risk")}
}
\description{
generate vaccination scenario according to input
}
