% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection_rules.R
\name{incremental}
\alias{incremental}
\title{routine coverage projection rule - annual incremental to certain threshold}
\usage{
incremental(
  d,
  year_from,
  year_to,
  step,
  cap = 0.95,
  gender = NULL,
  age_from = NULL,
  age_to = NULL
)
}
\arguments{
\item{d}{a data frame containing at least (year, coverage)}

\item{year_from}{year from}

\item{year_to}{year to}

\item{step}{incremental level, e.g. 1% annual increase}

\item{cap}{threshold capping coverage, this will be replaced by historical peak in d if applicable}

\item{gender}{1 for both, 2 for males, 3 for females}

\item{age_from}{age_from}

\item{age_to}{age_to}
}
\description{
routine coverage projection rule - annual incremental to certain threshold
}
