% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{vac_sce}
\alias{vac_sce}
\title{Generate Vaccination Scenario}
\usage{
vac_sce(input)
}
\arguments{
\item{input}{for details of setting up input, see input_check()}
}
\description{
generate vaccination scenario according to input
}
